Contracts/Creditworthiness.test.ts
describe('CreditWorthiness Contract', () => {
  let contract: CreditWorthiness; 
  let testAccount: Account; 

  beforeAll(async () => {
   
    contract = await deployContract<CreditWorthiness>(CreditWorthinessContract, testAccount);
  });

  it('should allow submitting an application', async () => {
    const applicationData = { /* Some test data representing an application */ };
    await expect(contract.submitApplication(applicationData, { from: testAccount })).resolves.not.toThrow();
  });

  it('should correctly verify an application', async () => {
    const applicationData = { /* Data that should pass verification */ };
    const result = await contract.verifyApplication(applicationData);
    expect(result).toBeTruthy();
  });

  it('should reject an invalid application', async () => {
    const invalidData = { /* Data that should fail verification */ };
    const result = await contract.verifyApplication(invalidData);
    expect(result).toBeFalsy();
  });

  // Add more test cases as needed
});
